Renamed a few files to match the current naming style
